Add dependency to build_fpm in make-images workflow
The 'build_nginx' job in the make-images GitHub Actions workflow has been updated to depend on the completion of 'build_fpm'. This ensures that the 'build_fpm' job finishes successfully before 'build_nginx' starts. The Docker 'build-args' is also updated to include 'FROM_FPM_IMAGE'. This is to ensure the nginx image is built from the appropriate fpm image.