Add GitHub workflow to update older image versions
This commit adds a new GitHub Workflow to the repository that runs on the first day of every month and whenever it is manually triggered in order to build and push outdated fpm and nginx images for PHP versions 5.6, 7.0, 7.1, 7.2, 7.3 & 7.4 to DockerHub.

It also removes PHP versions 7.2 and 7.4 from the existing workflow that builds images for the latest PHP versions i.e., 8.0, 8.1 and 8.2. The purpose of this separation is to optimize build time and resources.