Add multi-platform support to Github workflows and Makefile
Added support for different platforms including linux/amd64, linux/arm/v7, and linux/arm64/v8 in the Github workflows and Makefile.
This change ensures that images can be built and pushed for different processor architectures, expanding the reach and usability of our Docker images. The platforms are now defined in a matrix configuration, enabling easy expansion of supported platforms in the future.